requirements.txt
openai>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
pyyaml>=6.0
tenacity>=8.2.0
